meta-tq: linux-firmware: Do not apply ath10k dependency fix >=20240410
Commit 650e8837 ("ath10k: WCN3990: hw1.0: move firmware back from qcom/
location") moved the files to their original location. Beginning with
version 20240410 the file ath10k/WCN3990/hw1.0/wlanmdsp.mbn is no longer
a link but a regular firmware file which must not be deleted.
Limit the applicable version for the dependency workaround.

+++ meta-tq/recipes-kernel/linux-firmware/linux-firmware_%.bbappend
@@ -6,7 +6,7 @@ do_install:append() {
   # Remove backward-compatibility link from ath10k directory to qcom/sdm845.
   # This will create a pseudo dependency of linux-firmware-ath10k to linux-firmware-qcom-sdm845-modem.
   # Do not modify WHENCE file to not affect `WHENCE_CHKSUM`
-  if [ "${PV}" -ge "20231030" ]; then
+  if [ "${PV}" -ge "20231030" ] && [ "${PV}" -lt "20240410"]; then
     rm ${D}${nonarch_base_libdir}/firmware/ath10k/WCN3990/hw1.0/wlanmdsp.mbn
   fi
